requirements.txt
lxml 
pandas
psycopg2-binary
requests 
beautifulsoup4 
bs4 
SQLAlchemy
sqlalchemy_helpers
orm
future 
DateTime
pymongo